+++ gateway/gateway.go
@@ -330,9 +330,9 @@ type IPFSBackend interface {
 	// IsCached returns whether or not the path exists locally.
 	IsCached(context.Context, path.Path) bool
 
-	// GetIPNSRecord retrieves the best IPNS record for a given CID (libp2p-key)
+	// FindIPNS retrieves the best IPNS record for a given CID (libp2p-key)
 	// from the routing system.
-	GetIPNSRecord(context.Context, cid.Cid) ([]byte, error)
+	FindIPNS(context.Context, cid.Cid) ([]byte, error)
 
 	// ResolveMutable takes a mutable path and resolves it into an immutable one. This means recursively resolving any
 	// DNSLink or IPNS records.