+++ routing/http/contentrouter/contentrouter.go
@@ -20,10 +20,10 @@ import (
 var logger = logging.Logger("routing/http/contentrouter")
 
 type Client interface {
-	GetProviders(ctx context.Context, key cid.Cid) (iter.ResultIter[types.Record], error)
-	GetPeers(ctx context.Context, pid peer.ID) (peers iter.ResultIter[types.Record], err error)
-	GetIPNSRecord(ctx context.Context, name ipns.Name) (*ipns.Record, error)
-	PutIPNSRecord(ctx context.Context, name ipns.Name, record *ipns.Record) error
+	FindProviders(ctx context.Context, key cid.Cid) (iter.ResultIter[types.Record], error)
+	FindPeers(ctx context.Context, pid peer.ID) (peers iter.ResultIter[types.Record], err error)
+	FindIPNS(ctx context.Context, name ipns.Name) (*ipns.Record, error)
+	ProvideIPNS(ctx context.Context, name ipns.Name, record *ipns.Record) error
 }
 
 type contentRouter struct {
@@ -97,7 +97,7 @@ func readProviderResponses(iter iter.ResultIter[types.Record], ch chan<- peer.Ad
 }
 
 func (c *contentRouter) FindProvidersAsync(ctx context.Context, key cid.Cid, numResults int) <-chan peer.AddrInfo {
-	resultsIter, err := c.client.GetProviders(ctx, key)
+	resultsIter, err := c.client.FindProviders(ctx, key)
 	if err != nil {
 		logger.Warnw("error finding providers", "CID", key, "Error", err)
 		ch := make(chan peer.AddrInfo)
@@ -110,7 +110,7 @@ func (c *contentRouter) FindProvidersAsync(ctx context.Context, key cid.Cid, num
 }
 
 func (c *contentRouter) FindPeer(ctx context.Context, pid peer.ID) (peer.AddrInfo, error) {
-	iter, err := c.client.GetPeers(ctx, pid)
+	iter, err := c.client.FindPeers(ctx, pid)
 	if err != nil {
 		return peer.AddrInfo{}, err
 	}
@@ -164,7 +164,7 @@ func (c *contentRouter) PutValue(ctx context.Context, key string, data []byte, o
 		return err
 	}
 
-	return c.client.PutIPNSRecord(ctx, name, record)
+	return c.client.ProvideIPNS(ctx, name, record)
 }
 
 func (c *contentRouter) GetValue(ctx context.Context, key string, opts ...routing.Option) ([]byte, error) {
@@ -177,7 +177,7 @@ func (c *contentRouter) GetValue(ctx context.Context, key string, opts ...routin
 		return nil, err
 	}
 
-	record, err := c.client.GetIPNSRecord(ctx, name)
+	record, err := c.client.FindIPNS(ctx, name)
 	if err != nil {
 		return nil, err
 	}
@@ -198,7 +198,7 @@ func (c *contentRouter) SearchValue(ctx context.Context, key string, opts ...rou
 	ch := make(chan []byte)
 
 	go func() {
-		record, err := c.client.GetIPNSRecord(ctx, name)
+		record, err := c.client.FindIPNS(ctx, name)
 		if err != nil {
 			close(ch)
 			return